added meilisearch reset script

+++ apps/api/scripts/resetDevDatabase.js
@@ -40,6 +40,12 @@ readEnvFile().then((buffer) => {
 
         return datasource.query("SHOW TABLES WHERE Tables_in_soundcore NOT LIKE '%zone%' AND Tables_in_soundcore NOT LIKE '%mount%';").then(async (result) => {
             const tables = result.map((entry) => Object.values(entry)[0]);
+
+            if(tables.length <= 0) {
+                console.log("No tables deleted");
+                return;
+            }
+
             console.log(`Dropping ${tables.length} tables...`);
 
             const sql = "DROP TABLE " + tables.join(", ") + ";";
